requirements.txt
matplotlib==3.3.1
Flask==1.1.2
SQLAlchemy==1.3.19
Flask_SQLAlchemy==2.4.4
numpy==1.19.5
python-dotenv==0.19.0
